Update cpu_limit to float

+++ api/resource_usage/src/jupyverse_resource_usage/__init__.py
@@ -33,5 +33,5 @@ class ResourceUsageConfig(Config):
     mem_limit: int = 0
     mem_warning_threshold: float = 0.0
     track_cpu_percent: bool = False
-    cpu_limit: int = 0
+    cpu_limit: float = 0.0
     cpu_warning_threshold: float = 0.0
